Safe_Eye/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
ultralytics==8.1.27  # if you're using YOLO models

channels
daphne
djangorestframework-simplejwt

 ### --upgrade ultralytics
